Bug Fix: Asset Targets for CoreLib RootWorkbench not working properly
## Description
When a mod is created using the Workbench Definitions the mod itself will create and store the RootWorkbench info in each mod individually on the Asset Sprites. Causing 2 rootworkbench asset sprites with different IDs to be generated.

## Fix
On adding a new Workbench Definition, the target asset of the asset sprite skin is checked and changed to the proper rootworkbench asset sprite.